+++ src/lexi/tests.rs
@@ -309,7 +309,7 @@ fn write_lexiparser_source(replace_source: bool) {
         builder.print_items(0, false, false);
     }
     builder.set_indent(4);
-    let result_src = builder.gen_source_code();
+    let (result_src, ..) = builder.gen_source_code();
     if !cfg!(miri) {
         let expected_src = get_tagged_source(FILENAME, TAG).unwrap_or(String::new());
         if result_src != expected_src {